review(josheleonard): review fixes (part 1)
