+++ components/brave_wallet_ui/page/screens/swap/components/swap/select-token-modal/select-token-modal.tsx
@@ -195,13 +195,14 @@ export const SelectTokenModal = React.forwardRef<HTMLDivElement, Props>(
         </ScrollContainer>
         {showZeroBalanceButton && (
           <Button
-            buttonText={buttonText}
             onClick={toggleHideTokensWithZeroBalances}
             buttonStyle='square'
             buttonWidth='full'
             horizontalMargin={0}
             verticalMargin={0}
-          />
+          >
+            {buttonText}
+          </Button>
         )}
       </StandardModal>
     )
