+++ components/brave_wallet_ui/page/screens/swap/constants/types.ts
@@ -24,11 +24,14 @@ export type QuoteOption = {
   impact: Amount
   sources: LiquiditySource[]
 
-  // Indicates the kind of routing followed by the order.
-  // split -> indicates that the order was fulfilled from two separate LPs
-  //
-  // flow  -> indicates that the order was fulfilled through an intermediate
-  //          asset between two separate LPs.
+  /**
+   * Indicates the kind of routing followed by the order:
+   * 
+   * split -> indicates that the order was fulfilled from two separate LPs
+   * 
+   * flow  -> indicates that the order was fulfilled through an intermediate
+   *          asset between two separate LPs.
+   */
   routing: 'split' | 'flow'
 
   networkFee: string
@@ -99,9 +102,10 @@ export type SwapParams = {
   toToken?: BraveWallet.BlockchainToken
   fromAmount: string
   toAmount: string
-  // This is the value as seen on the UI -
-  // should be converted to appropriate
-  // format for Jupiter and 0x swap providers.
+  /**
+   * This is the value as seen on the UI - should be converted to appropriate 
+   * format for Jupiter and 0x swap providers.
+   **/
   slippageTolerance: string
   fromAddress?: string
   spotPrices: SpotPrices
